change a bunch of stuff into components
