+++ src/components/provider-search-result.tsx
@@ -3,15 +3,10 @@ import { Button, Col, Row } from 'react-bootstrap';
 import classNames from 'classnames';
 
 import SvgIcon from '@/components/svg-icon';
+import ProviderNextAppointmentCard, { SCHEDULE_TYPE_ID } from '@/components/provider-next-appointment-card';
 
 import { createUseThemedStyles } from '@/jss/theme';
 import mediaQueries from '@/jss/media-queries';
-
-export enum SCHEDULE_TYPE_ID {
-	APPOINTMENT_PREDETERMINED = 'APPOINTMENT_PREDETERMINED',
-	APPOINTMENT_UNDETERMINED = 'APPOINTMENT_UNDETERMINED',
-	APPOINTMENT_BY_PHONE = 'APPOINTMENT_BY_PHONE',
-}
 
 const useStyles = createUseThemedStyles((theme) => ({
 	providerResult: {
@@ -40,21 +35,6 @@ const useStyles = createUseThemedStyles((theme) => ({
 		'-webkit-box-orient': 'vertical',
 		'box-orient': 'vertical',
 		overflow: 'hidden',
-	},
-	scheduleCtaOuter: {
-		padding: 16,
-		borderRadius: 8,
-		boxShadow: theme.elevation.e200,
-		backgroundColor: theme.colors.n0,
-	},
-	iconOuter: {
-		width: 36,
-		height: 36,
-		display: 'flex',
-		borderRadius: 500,
-		alignItems: 'center',
-		justifyContent: 'center',
-		backgroundColor: theme.colors.p100,
 	},
 }));
 
@@ -114,48 +94,7 @@ const ProviderSearchResult = ({
 						<strong>Schedule Appointment</strong>
 					</p>
 					<p className="mb-4">{scheduleAppointmentDescription}</p>
-					{scheduleTypeId === SCHEDULE_TYPE_ID.APPOINTMENT_PREDETERMINED && (
-						<div className={classNames(classes.scheduleCtaOuter, 'mb-4')}>
-							<div className="d-md-flex justify-content-between">
-								<div className="mb-4 mb-md-0 me-4 d-flex align-items-center">
-									<div className={classNames(classes.iconOuter, 'me-4')}>
-										<SvgIcon kit="far" icon="calendar" size={16} className="text-primary" />
-									</div>
-									<div>
-										<p className="mb-0">First Available Appointment:</p>
-										<p className="mb-0">
-											<strong>Mon, May 4, 2:00PM</strong>
-										</p>
-									</div>
-								</div>
-								<Button variant="primary">Schedule Appointment</Button>
-							</div>
-						</div>
-					)}
-					{scheduleTypeId === SCHEDULE_TYPE_ID.APPOINTMENT_UNDETERMINED && (
-						<div className={classNames(classes.scheduleCtaOuter, 'mb-4')}>
-							<Button variant="primary" className="d-block w-100">
-								Schedule Appointment
-							</Button>
-						</div>
-					)}
-					{scheduleTypeId === SCHEDULE_TYPE_ID.APPOINTMENT_BY_PHONE && (
-						<div className={classNames(classes.scheduleCtaOuter, 'mb-4')}>
-							<div className="d-md-flex justify-content-between">
-								<div className="mb-4 mb-md-0 me-4 d-flex align-items-center">
-									<div className={classNames(classes.iconOuter, 'me-4')}>
-										<SvgIcon kit="far" icon="phone" size={16} className="text-primary" />
-									</div>
-									<div>
-										<p className="mb-0">
-											<strong>Call (000) 000-0000 to schedule</strong>
-										</p>
-									</div>
-								</div>
-								<Button variant="primary">Call Clinic</Button>
-							</div>
-						</div>
-					)}
+					<ProviderNextAppointmentCard scheduleTypeId={scheduleTypeId} />
 					<Button
 						variant="link"
 						className="d-block w-100 text-decoration-none"
